gauthd: supply the client singletons it links against
gauthd's Makefile linked gdeliveryd's gauthclient.o/gamedbclient.o objects
out of the box, which drags in the whole delivery daemon (state_GAuthClient,
GDeliveryServer, DomainDaemon, a dozen *::OnDBConnect handlers).  Instead the
two client managers gauthd references are now defined in gauthd itself:

  * GAuthClient::instance / GameDBClient::instance, so the RPC protocol
    manager can register their state tables at startup as before;
  * the out of line members gauthd's stubs and protocol code call
    (Reconnect, GetInitState, On*Session, OnCheckAddress, OnSetTransport,
    IdentifyFailed, SendProtocol) as thin forwarders to the connection
    manager, exactly like gdeliveryd's implementations.

gauthd never opens these two managers itself, so the forwarding stays inert
and callers answer ERR_DELIVER_SEND, which is what an authentication server
without a delivery session does today.  gauthd also gained an
attributes.size() cast in AccountingRequest::VerifyAuth so the OctetsStream
write keeps its 32 bit wire format on LP64.

+++ cnet/gauthd/accountingrequest.hpp
@@ -19,7 +19,7 @@ class AccountingRequest : public GNET::Protocol
 	bool need_response;
 	bool VerifyAuth()
 	{
-		return (MD5Hash(MD5HASH).Digest(OctetsStream()<<stamp<<userid<<attributes.size()<<GAuthServer::GetInstance()->shared_key) == authenticator);
+		return (MD5Hash(MD5HASH).Digest(OctetsStream()<<stamp<<userid<<(unsigned int)attributes.size()<<GAuthServer::GetInstance()->shared_key) == authenticator);
 	}
 	bool SendResponse(Manager::Session::ID sid)
 	{
